requirements.txt
pyscreenshot==3.1
numpy==1.26.2
sounddevice==0.4.6
pynput==1.7.6
opencv-python==4.8.1.78
PyYAML==5.4.1
psutil>=5.9.0
pywin32>=306; sys_platform == "win32"
pytesseract>=0.3.10; extra == "ocr"
Pillow>=10.0.0
